Move provider and facade into subfolders
