+++ composer.json
@@ -54,11 +54,11 @@
 		"laravel": {
 			"providers": [
 				"Aerni\\Spotify\\Providers\\SpotifyClientServiceProvider",
-				"Aerni\\Spotify\\SpotifyServiceProvider"
+				"Aerni\\Spotify\\Providers\\SpotifyServiceProvider"
 			],
 			"aliases": {
 				"SpotifyClient": "Aerni\\Spotify\\Facades\\SpotifyClientFacade",
-				"Spotify": "Aerni\\Spotify\\SpotifyFacade"
+				"Spotify": "Aerni\\Spotify\\Facades\\SpotifyFacade"
 			}
 		}
 	},
